Do HW6

+++ exercises/ex05/rand.h
@@ -8,7 +8,7 @@ float my_random_float();
   
 float my_random_float2();
   
-double my_random_double();
+float my_random_double();
   
 float dummy();
